Checkpoint. Going to experiment with 000_main.tf 001_vpc.tf 002_security_groups.tf 003_database.tf 005_parameter_store.tf 006_ec2.tf 007_load_balancer.tf 008_route53.tf 009_define_file_templates.tf 012_outputs.tf tterraform.tfvars variables.tf.
