+++ modules/connection_strings/v1.0.0/variables.tf
@@ -21,6 +21,11 @@ variable "flag_create_external_redis" {
   type        = bool
 }
 
+variable "flag_use_wave" {
+  description = "Whether to use Wave"
+  type        = bool
+}
+
 variable "flag_use_wave_lite" {
   description = "Whether to use Wave-Lite"
   type        = bool
