+++ tests/datafiles/external_redis_new.auto.tfvars
@@ -1,0 +1,3 @@
+
+flag_create_external_redis                      = true
+flag_use_container_redis                        = false
