+++ tests/datafiles/generate_override_data.sh
@@ -3,6 +3,9 @@
 echo "generate_override_data.sh: Current directory is $PWD"
 
 # Generate override data
+## ------------------------------------------------------------------------------------
+## New RDS
+## ------------------------------------------------------------------------------------
 echo "Generating external_db_new"
 cat << 'EOF' > external_db_new.auto.tfvars
 
@@ -12,6 +15,9 @@ flag_use_container_db                   = false
 EOF
 
 
+## ------------------------------------------------------------------------------------
+## Existing RDS
+## ------------------------------------------------------------------------------------
 echo "Generating external_db_existing"
 cat << 'EOF' > external_db_existing.auto.tfvars
 
@@ -20,4 +26,14 @@ flag_use_existing_external_db           = true
 flag_use_container_db                   = false
 
 tower_db_url =                          "mock-existing-tower-db.example.com"
+EOF
+
+## ------------------------------------------------------------------------------------
+## New Redis
+## ------------------------------------------------------------------------------------
+echo "Generating external_redis_new"
+cat << 'EOF' > external_redis_new.auto.tfvars
+
+flag_create_external_redis                      = true
+flag_use_container_redis                        = false
 EOF
