+++ tests/pytest.ini
@@ -1,0 +1,5 @@
+[pytest]
+markers =
+    local: Tests based on `terraform plan` only. Uses mocks for resources that would otherwise be created.
+    db:    Tests related to database resources.
+    redis: Tests related to Redis resources.
